# api.yaml
openapi: 3.0.2
info:
  title: DBS Programmierpraktikum - Phase 4 - API-Spezifikation
  description: >-
    In der folgenden Liste finden Sie die <b>für die Anwendergruppen
    spezifischen Grundfunktionalitäten</b>. Es müssen <b>alle folgenden
    allgemeinen Bedingungen</b> eingehalten werden: <ul>
      <li><b>Bei einer gültigen <code>GET</code>-Anfrage wird ein JSON-Array von JSON-Objekten, wobei ein JSON-Objekt eine Entität darstellt, zurückgegeben</b> und dies mit einem Code <code>200</code> belegt. Wird die <code>GET</code>-Anfrage auf einer nicht vorhandenen Ressource ausgeführt, so wird mit Code <code>404</code> geantwortet. Ist die Ressource woanders zu finden, wird dies mit dem Code <code>301</code> und dem Link im Header belegt. Eine <code>GET</code>-Anfrage auf einer Ressource, die eine Sammlung von Entitäten darstellt, kann durch Übergabe von bestimmten Parametern und dazugehörigen Werten <b>case-insensitive gefiltert</b> werden. Dabei muss jede Entität in der Sammlung allen Parametern gleichzeitig genügen. Die Sammlung wird in der Antwort in ein JSON-Array gekapselt.</li>
...
      <li><b>Bei einer gültigen <code>DELETE</code>-Anfrage wird der Code <code>204</code> zurückgegeben.</b> Ist die Ressource nicht vorhanden, wird dies mit Code <code>404</code> belegt.</li>
...
      <li><b>Antworten auf gültige <code>POST</code>- und <code>PATCH</code>-Anfragen beinhalten nichts.</b></li>
...
      <li><b>Passwörter werden der Einfachheit halber im Klartext übergeben.</b></li>
...
servers:
...
paths:
...
  /redakteure:
    get:
      tags:
        - Anwender
      summary: Gibt Redakteure zurück.
      description: >-
        Alle **Redakteure** werden zurückgegeben. Falls Query-Parameter mit Werten
        angegeben wurden, wird die Rückgabe noch entsprechend gefiltert.
      parameters:
        - name: nachname
          in: query
          description: 'Nur Redakteure, deren **Nachname** `nachname` enthält, werden zurückgegeben.'
          schema:
            type: string
      responses:
        '200':
          description: '**OK**.'
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  allOf:
                    - type: object
                      properties:
                        redakteurid:
                          type: integer
                          description: '**RowID des Redakteurs**.'
                        nutzerid:
                          type: integer
                          description: '**RowID des Nutzers**.'
                      required:
                        - redakteurid
                        - nutzerid
                    - $ref: '#/components/schemas/redakteur'
                    - $ref: '#/components/schemas/nutzer'
    post:
      tags:
        - Anwender
      summary: Fügt einen Redakteur hinzu.
      description: >-
        Ein **Redakteur** wird mit den entsprechenden Werten hinzugefügt. Ein zugeordneter Nutzer wird ebenfalls angelegt.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - $ref: '#/components/schemas/redakteur'
                - $ref: '#/components/schemas/nutzer'
      responses:
        '201':
          $ref: '#/components/responses/201'
        '400':
          $ref: '#/components/responses/400'
...
  /produktrezensionen:
    get:
      tags:
        - Anwender
      summary: Gibt Produktrezensionen zurück.
      description: >-
        Alle **Produktrezensionen** werden zurückgegeben. Falls Query-Parameter mit Werten angegeben wurden, wird die Rückgabe noch entsprechend gefiltert.
      parameters:
        - name: fazit
          in: query
          description: 'Nur Produktrezensionen, deren **Fazit** `fazit` enthält, werden zurückgegeben.'
          schema:
            type: string
        - name: beschreibung
          in: query
          description: 'Nur Produktrezensionen, deren **Produktbeschreibung** `beschreibung` enthält, werden zurückgegeben.'
          schema:
            type: string
      responses:
        '200':
          description: '**OK**.'
          content:
            application/json;charset=UTF-8:
              schema:
                type: array
                items:
                  allOf:
                    - type: object
                      properties:
                        blogeintragid:
                          type: integer
                          description: '**RowID des Blogeintrags**.'
                        produktrezensionid:
                          type: integer
                          description: '**RowID der Produktrezension**.'
                      required:
                        - produktrezensionid
                        - blogeintragid
                    - $ref: '#/components/schemas/produktrezension'
    post:
      tags:
        - Authentifizierter und als Redakteur authorisierter Anwender
      summary: Eine Produktrezension wird hinzugefügt.
      description: >-
        Eine **Produktrezension** wird hinzugefügt. Ein zugeordneter **Blogeintrag** wird ebenfalls angelegt.
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              allOf:
                - $ref: '#/components/schemas/produktrezension'
                - $ref: '#/components/schemas/blogeintrag'
      responses:
        '201':
          $ref: '#/components/responses/201'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'                    
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  responses:
    '201':
      description: '**Created**.'
      headers:
        Location:
          description: Die URL der neuen Ressource. Siehe hierzu auch die Anmerkung bzgl. fiktiver Links im ersten Abschnitt.
          schema:
            type: string
...
    '400':
      description: '**Bad Request**.'
    '401':
      description: '**Unauthorized**.'
      headers:
        WWW-Authenticate:
          schema:
            type: string
    '403':
      description: '**Forbidden**.'
...
  schemas:
    nutzer:
      type: object
      properties:
        email:
          type: string
          description: '**E-Mail-Adresse** des Nutzers.'
        passwort:
          type: string
          description: '**Passwort** des Nutzers.'
        geschlecht:
          type: string
          description: '**Geschlecht** des Nutzers.'
        benutzername:
          type: string
          description: '**Benutzername** des Nutzers.'
        geburtsdatum:
          type: string
          format: date
          description: '**Geburtsdatum** des Users im Format "YYYY-MM-DD".'
      required:
        - email
        - benutzername
        - passwort
        - geburtsdatum
        - geschlecht
    redakteur:
      type: object
      properties:
        vorname:
          type: string
          description: '**Vorname** des Redakteurs.'
        nachname:
          type: string
          description: '**Nachname** des Redakteurs.'
        biographie:
          type: string
          description: '**Biographie** des Redakteurs.'
      required:
        - vorname
        - nachname
...
    blogeintrag:
      type: object
      properties:
        erstellungsdatum:
          type: string
          format: date-time
          description: '**Erstellungsdatum** des Blogeintrags im Format "YYYY-MM-DD HH:MM:SS".'
        aenderungsdatum:
          type: string
          format: date-time
          description: '**Änderungsdatum** des Blogeintrags im Format "YYYY-MM-DD HH:MM:SS".'          
        text:
          type: string
          description: '**Text** des Blogeintrags.'
        titel:
          type: string
          description: '**Titel** des Blogeintrags.'
      required:
        - erstellungsdatum
        - text
        - titel
...
    produktrezension:
      type: object
      properties:
        beschreibung:
          type: string
          description: '**Beschreibung** der Produktrezension.'
        fazit:
          type: string
          description: '**Fazit** der Produktrezension.'
      required:
        - beschreibung
        - fazit
...